feat(tests): configure vitest for unit and browser testing
Add comprehensive Vitest configuration for running unit and browser tests.
Includes setup files, environment configurations, and coverage reporting.
