+++ front/.gitignore
@@ -11,8 +11,8 @@
 
 # production
 /build
-/junit-report.xml
-/sonar-report.xml
+junit-report.xml
+sonar-report.xml
 
 # misc
 .DS_Store
