+++ front/config/vitest/vitest.unit.config.ts
@@ -1,0 +1,19 @@
+import { resolve } from 'node:path';
+import { defineConfig } from 'vitest/config';
+import { getBaseConfig } from './base';
+
+export default defineConfig(({ mode }) => {
+  const base = getBaseConfig(mode);
+  return {
+    ...base,
+    test: {
+      ...base.test,
+      name: 'unit',
+      root: resolve(__dirname, '../../'),
+      environment: 'jsdom',
+      setupFiles: 'config/vitest/setup.unit.ts',
+      include: ['src/**/*.(spec|test|steps).[jt]s?(x)'],
+      exclude: ['src/**/*.browser.test.[jt]sx'],
+    },
+  };
+});
