+++ front/package.json
@@ -11,12 +11,12 @@
     "build": "rsbuild build",
     "build:storybook": "storybook build",
     "preview": "rsbuild preview",
-    "test": "vitest run --config vitest.workspace.ts",
-    "test:unit": "vitest run --config vitest.config.mts",
-    "test:unit:watch": "vitest --config vitest.config.mts",
-    "test:browser": "vitest --config=vitest.browser.config.ts --run",
-    "test:browser:update": "vitest --config=vitest.browser.config.ts --run --update",
-    "test:browser:watch": "vitest --config=vitest.browser.config.ts",
+    "test": "vitest run --config config/vitest/workspace.ts",
+    "test:unit": "vitest run --config config/vitest/vitest.unit.config.ts",
+    "test:unit:watch": "vitest --config config/vitest/vitest.unit.config.ts",
+    "test:browser": "vitest --config=config/vitest/vitest.browser.config.ts --run",
+    "test:browser:update": "vitest --config=config/vitest/vitest.browser.config.ts --run --update",
+    "test:browser:watch": "vitest --config=config/vitest/vitest.browser.config.ts",
     "lint": "node scripts/lint.js",
     "lint:oxlint": "oxlint src/ mocks/",
     "lint:oxlint:fix": "oxlint --fix src/ mocks/",
