+++ front/tsconfig.test.json
@@ -3,7 +3,6 @@
   "include": [
     "src",
     "mocks",
-    "vitest.setup.ts",
-    "vitest.browser.setup.ts",
+    "config/vitest",
   ],
 }
