--- front/vitest.config.mts
@@ -1,17 +1,0 @@
-import { defineConfig } from "vitest/config";
-import { getBaseConfig } from "./config/vitest.base";
-
-export default defineConfig(({ mode }) => {
-  const base = getBaseConfig(mode);
-  return {
-    ...base,
-    test: {
-      ...base.test,
-      name: 'unit',
-      environment: "jsdom",
-      setupFiles: "vitest.setup.ts",
-      include: ["src/**/*.(spec|test|steps).[jt]s?(x)"],
-      exclude: ["src/**/*.browser.test.[jt]sx"],
-    },
-  };
-});
